refactor(challenge): 거대 파일 3개 추가 분리 — TOP 3·4·5
MissionSubmitBonusSection (329줄 → 190줄)
- 비즈니스 로직 분리: mission-submit-bonus/useMissionSubmitBonus (188줄)
  - 상태·attendanceInfo 동기화·제출/수정/취소 mutation·약관 동의 등
- 약관 영역 추출: mission-submit-bonus/AgreementSection (56줄)
  - 개인정보 + 마케팅 두 약관이 동일 패턴이라 단일 컴포넌트로 묶음

MissionSubmitRegularSection (308줄 → 149줄)
- 비즈니스 로직 분리: mission-submit-regular/useMissionSubmitRegular (228줄)
  - 미션 종류 분기·attendanceInfo 동기화·보너스 미션 모달 트리거 등

BonusMissionSubmitMenu (302줄 → 199줄)
- 비즈니스 로직 분리: useBonusMissionSubmitMenu (138줄)
  - URL 정규식·링크 검증·계좌 입력·제출/수정 mutation 등

세 분리 모두 props/훅 인터페이스만 다듬고 동작 로직은 그대로 유지.
